% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_categorize.R
\name{age_categorize}
\alias{age_categorize}
\title{Categorize Age into Groups}
\usage{
age_categorize(df, col_name = NULL, method = NULL, var_name = NULL)
}
\arguments{
\item{df}{data frame containing an \code{age} variable.}

\item{col_name}{a string character of the \code{age} variable name in \code{df}.}

\item{method}{is a \code{numeric} code specifying the age groupin method. See \code{Details} for more information}
}
\value{
returns a \code{data frame} with new \code{factor} column of labels for the
specified age-groups.
}
\description{
\code{age_categorize} conveniently generates commonly used age groups from a continuous age variable
}
\section{Details}{

There are five pre-established age groups that can be specified using the \code{methods} parameter. 
Method \code{1}: 0, 1, ..., 119, 120+ (single year age groups)
Method \code{2}: <1, 0-4, ..., 70-74, 75+ (5 year age groups) 
Method \code{3}: 0-4, 5-9, ..., 70-74, 75+ (5 year age groups) 
Method \code{4}: <1, 1-4, ..., 95-99, 100+ (5 year age groups)
Method \code{5}: 0-4, 5-9, ..., 75-79, 80+ (UN demogratic standard)
Method \code{6}: 0-4, 5-9, 10-14, 15-59, 60+ (broad age groups)
Method \code{7}: 0-14, 15-59, 60+ (broad age groups)

The new column is generated using the \code{\link{mutate}} function from the \code{dplyr} package.
}

\examples{
## Use of the \\code{schizophrenia} dataset from the \\code{HSAUR} package
\dontrun{
require(HSAUR)
age_categorize(schizophrenia, col_name = "age", method = 5)
}

}
\seealso{
\code{\link{cut}}, \code{\link{dplyr::mutate}}
}
