% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtg_helpers.R
\name{attribute_as_name}
\alias{attribute_as_name}
\title{Attribute as name}
\usage{
attribute_as_name(attribute)
}
\arguments{
\item{attribute}{a quoted expression}
}
\value{
The attribute as a character
}
\description{
This function is used to get the attribute as a character instead of an expression
(e.g. attribute = Length) or an extract to the node (e.g. attribute = node$Length).
It is used in a special context where we know the attribute will always be used to reference
a node (i.e. data masking).
}
\examples{
\dontrun{
attribute_as_name(rlang::expr(Length))
attribute_as_name(rlang::expr(node$Length))
attribute_as_name(rlang::expr("Length"))

test = "Length"
attribute_as_name(rlang::expr(test))
}
}
\keyword{internal}
