% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplotr}
\alias{barplotr}
\title{Bar plot}
\usage{
barplotr(node, Property, y = NULL, MEAN = NULL, Interactive = NULL)
}
\arguments{
\item{node}{A data frame}

\item{Property}{A categorical column}

\item{y}{A numerical column. Specificy FALSE or node$columnname}

\item{MEAN}{Calculated average of numerical column known as y}

\item{Interactive}{An interactive html plot. Specificy TRUE or FALSE}
}
\value{
Barplots of Count, MEAN
}
\description{
Bar plot
}
