% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtg_helpers.R
\name{check_filters}
\alias{check_filters}
\title{Check filters}
\usage{
check_filters(node = NULL, scale = NULL, symbol = NULL, link = NULL)
}
\arguments{
\item{node}{The node}

\item{scale}{An integer vector of \code{.scale} to filter (should match the scales given by the SCALE
column from the MTG classes).}

\item{symbol}{A character vector of \code{.symbol} to filter (should match the symbols given by the SYMBOL
column from the MTG classes).}

\item{link}{A character vector of \code{.link} to filter}
}
\value{
Nothing. Return an error if anything went wrong.
}
\description{
Check the consistency of the filters
}
\examples{
\dontrun{
filepath= system.file("extdata", "simple_plant.mtg", package = "XploRer")
MTG = read_mtg(filepath)
check_filters(node = extract_node(MTG, "node_5"), scale = 1, symbol = "Individual",link = "/")
}
}
\keyword{internal}
