% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_df.R
\name{count_cat_df}
\alias{count_cat_df}
\title{Counts and Proportions for Categorical Variables}
\usage{
count_cat_df(df, ...)
}
\arguments{
\item{df}{a \code{data frame} object to report descriptives on}

\item{...}{names of categorical variables to report the descriptives by group}
}
\value{
Returns a \code{data frame} with total counts by group and percentage of total by group
}
\description{
This function returns descriptives for categorical variables 
(reports total counts (n) and percentage of total)
}
\section{Warning}{

If columns with non-continuous values (numeric or integer) exist in \code{df}, an error 
is thrown and returns potential columns that you will have to remove
for the function to work.
}

\examples{
iris \%>\% 
 mutate(Bin = sample(x = 1:3,size = 150,replace = TRUE)) \%>\% 
 count_cat_df(Species, Bin)

}
