% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose_}
\alias{decompose_}
\title{Get decomposing nodes recursively}
\usage{
decompose_(
  attribute,
  node = NULL,
  ref_node = NULL,
  scale = NULL,
  symbol = NULL,
  link = NULL,
  filter_fun = NULL,
  decomp_type = c("scale", "symbol")
)
}
\arguments{
\item{attribute}{Any node attribute (as a character)}

\item{node}{The MTG node}

\item{ref_node}{The reference node to match the scale or symbol}

\item{scale}{An integer vector for filtering the \code{.scale} of the decomposing nodes (i.e. the SCALE
column from the MTG classes).}

\item{symbol}{A character vector for filtering the decomposing nodes by the name of their \code{.symbol} (i.e. the SYMBOL
column from the MTG classes).}

\item{link}{A character vector for filtering the \code{.link} with the descendant}

\item{filter_fun}{Any filtering function taking a node as input.}

\item{decomp_type}{The type of scale to use for decomposition, either the scale or the symbol of the node.}
}
\value{
The attributes of the nodes decomposing the reference node.
}
\description{
This is the function actually doing the hard work for \verb{[decompose()]}.
}
\keyword{internal}
