% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_df.R
\name{describe_df}
\alias{describe_df}
\title{Summarizes Missing Data, Data Types, and Foreign Labels into a tidy data frame}
\usage{
describe_df(df, vars = names(df), ...)
}
\arguments{
\item{df}{a \code{data frame} object to identify the foreign labels from select variables of a data frame}

\item{vars}{a \code{character string} or \code{character vector} of column names in the data frame to examine}
}
\value{
returns a \code{data frame} where column from \code{df} is a row and
information on the number of missing values, data type, and foreign labels are specified.
}
\description{
This function uses \code{type_df}, \code{missing_df}, \code{lab_foreign_df} to provide information on the
variables contained within the \code{df} input.
}
\examples{
describe_df(airquality)

}
