% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_df.R
\name{describe_viz}
\alias{describe_viz}
\title{Plots Missing Data, Data Types, and Foreign Labels into a tidy data frame}
\usage{
describe_viz(df, title = NULL, subtitle = NULL, threshold = 10, ...)
}
\arguments{
\item{df}{a \code{data frame} object to identify the foreign labels from select variables of a data frame}

\item{vars}{a \code{character string} or \code{character vector} of column names in the data frame to examine}
}
\value{
returns a \code{ggplot2} plot displaying the percentage of missing values and data types
for each column contained in the data frame
}
\description{
This function takes the \code{describe_df} function, built on
\code{type_df}, \code{missing_df}, \code{lab_foreign_df} to plot information on the
variables contained within the \code{df} input.
}
\examples{
describe_df(airquality)
describe_viz(airquality)

}
