% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsos.R
\name{.ls.objects}
\alias{.ls.objects}
\title{List Objects in Global Environment}
\usage{
.ls.objects(pos = 1, pattern, order.by, decreasing = FALSE,
  head = FALSE, n = 5)
}
\description{
This \code{lsos} function summarizes all objects found in the global environment into a \code{data frame}.
If using \code{RStudio}, this information can be viewed using the "Environment" Tab in the GUI.
}
\details{
See this Stack Overflow post: https://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session
}
\examples{
data(iris)
data(airquality)

x <- letters[1:15]
y <- matrix(data = rnorm(n = 50),ncol = 10)

lsos(n = 50)

}
