% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MTG.R
\name{extends_point}
\alias{extends_point}
\title{Extends coordinate}
\usage{
extends_point(x0, y0, x1, y1, extend_length)
}
\arguments{
\item{x0}{x coordinate of the origin}

\item{y0}{y coordinate of the origin}

\item{x1}{x coordinate of the point}

\item{y1}{y coordinate of the point}

\item{extend_length}{ROtation angle in degree (anti-clockwise)}
}
\value{
The new coordinates for the point
}
\description{
Extends coordinate
}
\examples{
\dontrun{
extends_point(0,0,1,0,90)
}
}
\keyword{internal}
