% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtg_helpers.R
\name{extract_node}
\alias{extract_node}
\title{Extract a node from the MTG}
\usage{
extract_node(mtg, name)
}
\arguments{
\item{mtg}{An mtg as from \code{\link[=read_mtg]{read_mtg()}}}

\item{name}{The name of the node, e.g. "node_1"}
}
\value{
The node (with all its information, parent, children, attributes...)
}
\description{
Extract a node from the MTG
}
\examples{
filepath= system.file("extdata", "tree1h.mtg", package = "XploRer")
MTG = read_mtg(filepath)
extract_node(MTG,"node_6")
}
