% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_MTG.R
\name{find_MTG_node_column}
\alias{find_MTG_node_column}
\title{Find MTG node column}
\usage{
find_MTG_node_column(MTG_node)
}
\arguments{
\item{MTG_node}{An MTG node (e.g. "/Individual0")}
}
\value{
A parsed node in the form of a list of three:
\itemize{
\item the link
\item the name
\item and the index
}
}
\description{
Find the position of a node in the MTG (which column it belongs to)
}
\examples{
\dontrun{
# First column (last column is missing):
find_MTG_node_column(c("/P1","/A1",""))
# Node starts at column 2:
find_MTG_node_column(c("","/U1","<U2"))
}
}
\keyword{internal}
