% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtg_helpers.R
\name{get_follow_values}
\alias{get_follow_values}
\title{Get consecutive node values}
\usage{
get_follow_values(attribute, node = NULL, scale = NULL, symbol = NULL)
}
\arguments{
\item{attribute}{Any node attribute (as a character)}

\item{node}{The MTG node}

\item{symbol}{A character vector for filtering the names of the nodes that are considered for succession.}
}
\value{
A vector of values named after the nodes of interest
}
\description{
Get the values of an attribute for all consecutive nodes (i.e. following nodes,
excluding branching ones). It is usually used to compute e.g. the total length
of an axis when the length were measured at the segment scale or unit of growth scale.
}
\examples{
filepath= system.file("extdata", "tree1h.mtg", package = "XploRer")
MTG = read_mtg(filepath)
get_follow_values(attribute = "length", node = extract_node(MTG,"node_8"), symbol = "S")

# Length were observed at the "S" scale (S = segment of an axis between tow branches),
# but we need the length at the axis scale, to do so:
mutate_mtg(MTG, axis_length = sum(get_follow_values(attribute = "length", symbol = "S")), .symbol = "A")
}
