% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtg_helpers.R
\name{get_nodes_symbol_values}
\alias{get_nodes_symbol_values}
\title{Get consecutive node values}
\usage{
get_nodes_symbol_values(attribute, node = NULL, symbol = NULL)
}
\arguments{
\item{attribute}{Any node attribute (as a character)}

\item{node}{The MTG node}

\item{symbol}{A character vector for filtering the names of the nodes that are considered for succession.}
}
\value{
A vector of values named after the nodes of interest
}
\description{
Get the values of an attribute for all consecutive nodes (i.e. not branching)
}
\examples{
filepath= system.file("extdata", "tree1h.mtg", package = "XploRer")
MTG = read_mtg(filepath)
get_nodes_symbol_values(attribute = "length", node = FindNode(MTG$MTG,"node_8"), symbol = "S")

}
