% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtg_apply.R
\name{get_parent_value}
\alias{get_parent_value}
\title{Get parent value}
\usage{
get_parent_value(..., node, .scale = NULL, .type = NULL)
}
\arguments{
\item{...}{Any node attribute (as a character)}

\item{node}{The node (do not put something when used from \code{\link[=mtg_apply]{mtg_apply()}})}

\item{.scale}{The names of the MTG scale required (i.e. the SYMBOL from the MTG classes)}
}
\value{
The values of the attributes for the parent of the node
}
\description{
Get the value of one or several variables from the parent node. This is a helper
function that is usually used as input to \code{\link[=mtg_apply]{mtg_apply()}} to get the values of the parent node
for all nodes.
}
\details{
This function returns the values of any attribute of the parent of a node. It is
mainly intended to be used in a call to \code{\link[=mtg_apply]{mtg_apply()}} (see examples).
}
\examples{
filepath= system.file("extdata", "simple_plant.mtg", package = "XploRer")
MTG = read_MTG(filepath)
get_parent_value("Length","Width",  node = FindNode(MTG$MTG, "node_5"))
}
