% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_df.R
\name{lab_foreign_df}
\alias{lab_foreign_df}
\title{Identifies foreign labels within a data frame}
\usage{
lab_foreign_df(df, vars = names(df))
}
\arguments{
\item{df}{a \code{data frame} object to identify the foreign labels from select variables of a data frame}

\item{vars}{a \code{character string} or \code{character vector} of column names in the data frame to examine}
}
\value{
returns a \code{data frame} with a \code{logical}  column containing whether the specified
columns have a foreign label associated it.
}
\description{
This function identifies labels associated with foreign data types such as Stata or SPSS.
Typically object types are imported using the \code{\link{haven}} package.
}
\examples{
type_df(airquality)
lab_foreign_df(airquality)
}
