% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtg_getters_shortcuts.R
\name{leaves}
\alias{leaves}
\title{Get leaves values}
\usage{
leaves(
  attribute,
  node = NULL,
  scale = NULL,
  symbol = NULL,
  link = NULL,
  filter_fun = NULL
)
}
\arguments{
\item{attribute}{Any node attribute (as a character)}

\item{node}{The MTG node}

\item{scale}{An integer vector for filtering the \code{.scale} of the children (i.e. the SCALE
column from the MTG classes).}

\item{symbol}{A character vector for filtering the children by the name of their \code{.symbol} (i.e. the SYMBOL
column from the MTG classes).}

\item{link}{A character vector for filtering the \code{.link} with the children}

\item{filter_fun}{Any filtering function taking a node as input.}
}
\value{
The attribute values from the children of the node
}
\description{
Get attribute values from the terminal nodes of the descendants sub-tree of a
node (i.e. the "leaves" in computer science terms).
}
\note{
This function is mainly intended to be used with \code{\link[=mutate_mtg]{mutate_mtg()}}. In this case,
the \code{node} argument can be left empty (or you can put \code{node = node} equivalently).
}
\examples{
filepath= system.file("extdata", "simple_plant.mtg", package = "XploRer")
MTG = read_mtg(filepath)

# The MTG has two leaves (node_5 and node_7):
leaves("Length", node = MTG)
# We can check the function worked well by printing the symbols of each nodes:
print(MTG, ".symbol")

# If we need the terminal Internodes:
leaves("Length", node = MTG, symbol = "Internode")

# Or the leaves at a given scale:
print(MTG, ".scale")
leaves("Length", node = MTG, scale = 2)

}
