% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_df.R
\name{missing_df}
\alias{missing_df}
\title{Quantifies missing values found within a data frame}
\usage{
missing_df(df, vars = names(df))
}
\arguments{
\item{df}{a \code{data frame} object to identify missing values}

\item{vars}{a \code{character string} or \code{character vector} of column names in the data frame to examine}
}
\value{
Returns a \code{data frame} with four columns, each row displays the number and percentage of missing
values, and the percentage of existing (non-missing) values for each variable within the data frame
}
\description{
This \code{map} wrapper function vectorizes the count and proportion of missing and non-missing values within
a tidy data frame.
}
\examples{
missing_df(airquality)

}
