% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_jenks_brks.R
\name{mutate_jenks_brks}
\alias{mutate_jenks_brks}
\title{Generate Jenks Breaks for Mapping with ggplot2}
\usage{
mutate_jenks_brks(input, variable, ...)
}
\arguments{
\item{input}{a `data.frame` or `spatial` object}

\item{variable}{name (using standard evaluation) a continuous variable either `integer` or `numeric` to apply Jenks breaks}

\item{...}{arguments used in `classIntervals` specific to `style = "jenks"`}
}
\value{
returns the `input` object with a new `factor` columns named as `variable_brks` that contains 
Jenks breaks and corresponding labels for each `factor` level
}
\description{
This function is a wrapper around the `dplyr` `mutate()` function that generates Jenks breaks using the `classIntervals` package.
See \code{\link{classIntervals}} for more details on methods.
}
\section{Warning}{

The Jenks breaks do not like missing values and will give a warning message
when the variable of interest contains missing values!
}

\examples{
#Apply Jenks breaks on the Solar Radiation column frmo the airqualtiy dataset
airquality \%>\% mutate_jenks_brks(Solar.R)

\dontrun{
# Ebola Cause of Death joined to Chiefdom Shapefiles
# create Jenks breaks for deaths with 6 breaks

evd_shp <- evd_shp \%>\% mutate_jenks_brks(deaths, n = 6)

p_evd_deaths <- ggplot(evd_shp) + geom_sf(aes(fill=deaths_brks)) +
scale_fill_brewer(type = "seq", palette = 9, name = "Deaths") +
theme_void() 
}
}
