% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_labs.R
\name{mutate_lab_foreign}
\alias{mutate_lab_foreign}
\title{Mutate Wrapper for Foreign Labels}
\usage{
mutate_lab_foreign(df, col_name = NULL, regex = NULL)
}
\arguments{
\item{df}{a \code{data frame} object to identify missing values}

\item{regex}{a \code{character} string containing a regular expression to match a column or columns}

\item{column_name}{the name of the specified column (uses standard evalulation typical of \code{dplyr})}
}
\value{
Returns a \code{data frame} with a newly appended column
contain the respective foreign labels from the \code{haven} package for each row
based on the specified column.
}
\description{
This \code{mutate_lab_foreign} wrapper function uses the \code{\link{mutate}} 
function from the \code{dplyr} to create a new column containing the foreign
labels from a selected column.
}
\examples{
mutate_lab_foreign(airquality, Solar.R) #throws error because no columns contain foreign labels

}
