% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_MTG.R
\name{parse_MTG_classes}
\alias{parse_MTG_classes}
\title{Parse MTG classes}
\usage{
parse_MTG_classes(MTG)
}
\arguments{
\item{MTG}{A pre-formatted MTG}
}
\value{
A data.frame with the classes
}
\description{
Parse the classes section from an MTG
}
\keyword{internal}
