% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_MTG.R
\name{parse_MTG_code}
\alias{parse_MTG_code}
\title{Parse MTG code}
\usage{
parse_MTG_code(MTG)
}
\arguments{
\item{MTG}{A pre-formatted MTG}
}
\value{
The MTG code
}
\description{
Parse the code section from an MTG
}
\keyword{internal}
