% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_MTG.R
\name{parse_MTG_description}
\alias{parse_MTG_description}
\title{Parse MTG description}
\usage{
parse_MTG_description(MTG)
}
\arguments{
\item{MTG}{A pre-formatted MTG}
}
\value{
A data.frame with the description
}
\description{
Parse the description section from an MTG
}
\keyword{internal}
