% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_MTG.R
\name{parse_MTG_features}
\alias{parse_MTG_features}
\title{Parse MTG features}
\usage{
parse_MTG_features(MTG)
}
\arguments{
\item{MTG}{A pre-formatted MTG}
}
\value{
A data.frame with the features
}
\description{
Parse the features section from an MTG
}
\keyword{internal}
