% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_MTG.R
\name{parse_MTG_lines}
\alias{parse_MTG_lines}
\title{Parse MTG lines}
\usage{
parse_MTG_lines(MTG_code, classes, description, features)
}
\arguments{
\item{classes}{The MTG classes}

\item{description}{The MTG description}

\item{features}{The MTG features}

\item{MTG}{A pre-formatted MTG}
}
\value{
A parsed MTG
}
\description{
Parse the MTG lines (called from \code{\link[=parse_MTG_MTG]{parse_MTG_MTG()}})
}
\keyword{internal}
