% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_MTG.R
\name{parse_MTG_MTG}
\alias{parse_MTG_MTG}
\title{Parse MTG}
\usage{
parse_MTG_MTG(MTG, classes, description, features, first_line = 1)
}
\arguments{
\item{MTG}{A pre-formatted MTG}

\item{classes}{The MTG classes}

\item{description}{The MTG description}

\item{features}{The MTG features}

\item{first_line}{The first line of the MTG. Used to print the true line at
which there is an error.}
}
\value{
A parsed MTG
}
\description{
Parse the MTG section from an MTG file
}
\keyword{internal}
