% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pieplot.R
\name{pieplotr}
\alias{pieplotr}
\title{Pie plot and nested pie plot}
\usage{
pieplotr(node, ..., ProjectID = NULL)
}
\arguments{
\item{node}{A dataframe}

\item{ProjectID}{Specify TRUE or FALSE to include Project ID into pieplot}

\item{Property}{A categorical column}
}
\description{
Pie plot and nested pie plot
}
