% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MTG.R
\name{plot.mtg}
\alias{plot.mtg}
\title{Plot an MTG}
\usage{
\method{plot}{mtg}(..., scale = NULL, angle = 45, phylotaxy = TRUE)
}
\arguments{
\item{...}{An MTG, as from \code{\link[=read_MTG]{read_MTG()}}}

\item{scale}{The scale required for plotting}

\item{angle}{Insertion angle when branching}

\item{phylotaxy}{Is phylotaxy required ? Uses 180 degrees if \code{TRUE}.}
}
\value{
A ggplot of the MTG
}
\description{
Plot an MTG
}
\details{
The function needs the topological orders as attributes in the MTG. If they
are not present, it uses \code{\link[=topological_order]{topological_order()}} to compute it with descending order.
If you need ascending order, please use \code{topological_order(MTG, ascend = TRUE)} before
plotting.
}
\examples{
filepath= system.file("extdata", "simple_plant.mtg", package = "XploRer")
MTG= read_MTG(filepath)
plot(MTG)
}
