% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MTG.R
\name{plot_MTG}
\alias{plot_MTG}
\title{Plot an MTG}
\usage{
plot_MTG(MTG, scale = NULL)
}
\arguments{
\item{MTG}{An MTG, as from \code{\link[=read_MTG]{read_MTG()}}}

\item{scale}{The scale required for plotting}
}
\value{
A ggplot of the MTG
}
\description{
Plot an MTG
}
\examples{
filepath= system.file("extdata", "simple_plant.mtg", package = "XploRer")
MTG= read_MTG(filepath)
plot(MTG)

}
