% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MTG.R
\name{plotly_MTG}
\alias{plotly_MTG}
\title{Plot an interactive MTG}
\usage{
plotly_MTG(MTG, scale = NULL, angle = 45, phylotaxy = TRUE, ...)
}
\arguments{
\item{MTG}{An MTG, as from \code{\link[=read_MTG]{read_MTG()}}}

\item{scale}{The scale required for plotting}

\item{angle}{Insertion angle when branching}

\item{phylotaxy}{Is phylotaxy required ? Uses 180 degrees if \code{TRUE}.}

\item{...}{Names of the variables to be added to the tooltip.}
}
\value{
A \link{plotly} object of the MTG
}
\description{
Plot an interactive MTG
}
\examples{
filepath= system.file("extdata", "simple_plant.mtg", package = "XploRer")
MTG= read_MTG(filepath)
plotly_MTG(MTG)
}
