% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_MTG.R
\name{read_MTG}
\alias{read_MTG}
\title{Read MTG}
\usage{
read_MTG(file)
}
\arguments{
\item{file}{The path to the MTG file}
}
\value{
A named list of four sections: the MTG classes, description, features,
and MTG. The MTG is a \link{data.tree} data structure.
}
\description{
Read an MTG file.
}
\note{
See the documentation for the MTG format from the
\href{http://openalea.gforge.inria.fr/doc/vplants/newmtg/doc/_build/html/user/intro.html#mtg-a-plant-architecture-databases}{OpenAlea webpage}
for further details.
}
\examples{
\dontrun{
filepath= system.file("extdata", "simple_plant.mtg", package = "XploRer")
read_MTG(filepath)
}
}
