% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_MTG.R
\name{read_MTG_classes}
\alias{read_MTG_classes}
\title{Read MTG classes}
\usage{
read_MTG_classes(MTG)
}
\arguments{
\item{MTG}{A pre-formatted MTG}
}
\value{
A data.frame with the classes
}
\description{
Read the classes section from an MTG
}
\keyword{internal}
