% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_MTG.R
\name{read_MTG_code}
\alias{read_MTG_code}
\title{Read MTG code}
\usage{
read_MTG_code(MTG)
}
\arguments{
\item{MTG}{A pre-formatted MTG}
}
\value{
The MTG code
}
\description{
Read the code section from an MTG
}
\keyword{internal}
