% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_MTG.R
\name{read_MTG_description}
\alias{read_MTG_description}
\title{Read MTG description}
\usage{
read_MTG_description(MTG)
}
\arguments{
\item{MTG}{A pre-formatted MTG}
}
\value{
A data.frame with the description
}
\description{
Read the description section from an MTG
}
\keyword{internal}
