% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_MTG.R
\name{read_MTG_section}
\alias{read_MTG_section}
\title{Read MTG section}
\usage{
read_MTG_section(MTG, section_name, header, next_section, allow_empty)
}
\arguments{
\item{MTG}{A pre-formatted MTG}

\item{section_name}{The section name in the file}

\item{header}{The header of the section}

\item{next_section}{The name of the next section}

\item{allow_empty}{Boolean. Allow the section to be empty ?}
}
\value{
A data.frame with the section results, or a tree for the MTG.
}
\description{
Read any section from an MTG
}
\examples{
\dontrun{
filepath= system.file("extdata", "simple_plant.mtg", package = "XploRer")
MTG_file = strip_comments(MTG_file)
MTG_file = strip_empty_lines(MTG_file)

# Checking that all sections are present and ordered properly:
check_sections(MTG_file)
# Read the classes section from the MTG:
read_MTG_section(MTG_file, "CLASSES:",
  c('SYMBOL', 'SCALE', 'DECOMPOSITION', 'INDEXATION', 'DEFINITION'),
  "DESCRIPTION:",FALSE)
}
}
\keyword{internal}
