% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MTG.R
\name{rotate_point}
\alias{rotate_point}
\title{Rotate coordinates}
\usage{
rotate_point(x0, y0, x1, y1, angle)
}
\arguments{
\item{x0}{x coordinate of the origin}

\item{y0}{y coordinate of the origin}

\item{x1}{x coordinate of the point}

\item{y1}{y coordinate of the point}

\item{angle}{Rotation angle in degree (anti-clockwise)}
}
\value{
The new coordinates for the point
}
\description{
Rotate coordinates
}
\examples{
rotate_point(0,0,1,0,90)

# Example with a plot:
x0 = 1 ; y0 = 1 ; x1= 2 ; y1= 2 ;x3= 3 ; y3= 3
plot(c(x0,x1),c(y0,y1), type="l", ylim = c(0,3), xlim = c(0,3))
segments(x1,y1,x3,y3, col= 2)
segments(x1,y1,rotate_point(x1,y1,x3,y3, 45)[1],rotate_point(x1,y1,x3,y3, 45)[2], col= 3)
}
\keyword{internal}
