% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_all.R
\name{see_all}
\alias{see_all}
\title{Display Variable Details for all variables in a data.frame}
\usage{
see_all(x)
}
\arguments{
\item{x}{The data.frame to be displayed in the console.}
}
\description{
This function displays the following details of each variable:
\itemize{
 \item{attributes (i.e. useful to inspect labels)}
 \item{class, typeof, mode, storage.mode}
 \item{table (frequency and percentage of values)}
 \item{summary statistics}
}
See is an S3 generic and currently works with objects of class 'character', 'factor', 'numeric', 'data.frame', 'labelled', 'haven_labelled', 'POSIXct' and 'Date'.
}
\examples{
see_all(mtcars)
}
\keyword{codebook}
\keyword{see}
