% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_at.R
\name{see_at}
\alias{see_at}
\title{Display Variable Details for a selction of variables in a data.frame}
\usage{
see_at(data, variables)
}
\arguments{
\item{data}{The data.frame from which variables are selected.}

\item{variables}{The variables that should be displayed, in quotes.}
}
\description{
This function displays the following details of each variable:
\itemize{
 \item{attributes (i.e. useful to inspect labels)}
 \item{class, typeof, mode, storage.mode}
 \item{table (frequency and percentage of values)}
 \item{summary statistics}
}
See is an S3 generic and currently works with objects of class 'character', 'factor', 'numeric', 'data.frame', 'labelled', 'haven_labelled', 'POSIXct' and 'Date'.
}
\examples{
see_at(mtcars, variables = c("cyl", "mpg"))
}
\keyword{codebook}
\keyword{see}
