% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-string.R
\name{split_at_blank}
\alias{split_at_blank}
\title{Split string at blank}
\usage{
split_at_blank(str)
}
\arguments{
\item{str}{A character string}
}
\value{
A vector of strings
}
\description{
Split a string using blank characters (i.e. one or more tab or space)
}
\examples{
split_at_blank(c("word1 word2   word3  word4"))
# NB: first separation uses one space, second two spaces, third a tabulation
}
