% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-string.R
\name{split_at_tab}
\alias{split_at_tab}
\title{Split string at tab}
\usage{
split_at_tab(str)
}
\arguments{
\item{str}{A character string}
}
\value{
A vector of strings
}
\description{
Split a string using tabulation characters
}
\examples{
\dontrun{
split_at_tab(c("word1\tword2"))
}
}
\keyword{internal}
