% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_MTG.R
\name{split_MTG_elements}
\alias{split_MTG_elements}
\title{Split MTG line}
\usage{
split_MTG_elements(MTG_line)
}
\arguments{
\item{MTG_line}{An MTG line (e.g. "/P1/A1")}
}
\value{
A vector of elements (keeping their link, e.g. + or <)
}
\description{
Split the elements (e.g. inter-node, growth unit...) in an MTG line
}
\examples{
\donttest{
split_MTG_elements("/A1+U85/U86<U87<.<U93<U94<.<U96<U97+.+U100")
}
}
\keyword{internal}
