% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss_col_labels.R
\name{spss_code_labels}
\alias{spss_code_labels}
\title{Extract Code Labels for a Foreign Labelled Column}
\usage{
spss_code_labels(df, vars = NULL)
}
\arguments{
\item{df}{a \code{data frame} object to identify missing values}

\item{vars}{a column name (uses standard evalauation) in the data frame to examine any foreign labels}
}
\value{
a \code{data frame} containing each code or label associated with each unique value (or level) 
found in the specific column. 
If there are no foreign labels, then a zero-dimnensional \code{tibble} is returned
}
\description{
This function extracts any foreign code labels of a specified column from a \code{data frame} input. 
This is analagous to the \code{\link{levels}} function that extracts the labels or levels
from a \code{factor} data type.
}
\examples{
spss_code_labels(airquality, Month)

}
