% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss_col_labels.R
\name{spss_col_labels_vec}
\alias{spss_col_labels_vec}
\title{Extracts Column Labels from a Foreign Data Frame}
\usage{
spss_col_labels_vec(df, col_name = NULL)
}
\arguments{
\item{df}{a \code{data frame} object to identify missing values}
}
\value{
Returns a \code{characer} vector with each column from the \code{data frame} input as a row and 
the corresponding foreign label (SPSS, Stata, SAS) for any applicable columns. 
Any columns with no foreign labels return \code{NA} values for the \code{column_labels} column.
}
\description{
This function extracts any foreign column labels from a \code{data frame} input. It 
searches the \code{attributes} of the \code{data frame} input for named values  as \code{$label}.
}
\examples{
spss_col_labels(airquality)

}
