% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_MTG.R
\name{strip_comments}
\alias{strip_comments}
\title{Remove comments}
\usage{
strip_comments(str)
}
\arguments{
\item{str}{A character string}
}
\value{
The character string stripped from its comments
}
\description{
Remove comments from a string
}
\examples{
strip_comments(c("test","test2 # with a comment", "# just a comment"))
}
