% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_df.R
\name{summarise_df}
\alias{summarise_df}
\title{Summary Statistics for Continuous Variables}
\usage{
summarise_df(df, vars = names(df))
}
\arguments{
\item{df}{a \code{data frame} object to report summary statistic values}

\item{vars}{is a \code{character} string (single length or not) of the variables of interest. 
The default value is it selects all column in \code{df}.}
}
\value{
Returns a \code{data frame} where each row is a column from the original
data frame and each column is a summary statistic.
}
\description{
This function returns summary statistics 
(mean, median, minimum, maximum, standard deviation, counts, and missing values) for
continuous variables found in a datamfrae
}
\section{Warning}{

If columns with non-continuous values (numeric or integer) exist in \code{df}, an error 
is thrown and returns potential columns that you will have to remove
for the function to work.
}

\examples{
summarise_df(airquality)

}
