% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{tabler}
\alias{tabler}
\title{Table}
\usage{
tabler(node, Property, ProjectID = NULL, y = NULL, MEAN = NULL)
}
\arguments{
\item{node}{A dataframe}

\item{ProjectID}{Specify TRUE or FALSE to include Project ID in count}

\item{y}{A numerical column. Specify FALSE or node$columnname}

\item{MEAN}{Calculated average of numerical column known as y}

\item{Property_Name}{A categorical column}

\item{UniqueProjectID}{Specify TRUE or FALSE to count number of Unique Project IDs - CURRENTLY DEPRACATED}
}
\description{
Table
}
