% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological_order.R
\name{topological_order}
\alias{topological_order}
\title{Topological order}
\usage{
topological_order(MTG, ascend = FALSE)
}
\arguments{
\item{MTG}{An MTG tree as from \code{\link[=read_mtg]{read_mtg()}}}

\item{ascend}{Is the order computed from the base (\code{TRUE}), or the tip (\code{FALSE}) ?}
}
\value{
Nothing, update the values of the MTG in-place. The MTG is enriched with the
topological order of each node
}
\description{
Compute the topological order of a tree
}
\examples{
filepath= system.file("extdata", "simple_plant.mtg", package = "XploRer")
MTG= read_mtg(filepath)
topological_order(MTG)
print(MTG, "topological_order")
}
