% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_df.R
\name{type_df}
\alias{type_df}
\title{Identifies class for each variable in data frame}
\usage{
type_df(df, vars = names(df))
}
\arguments{
\item{df}{a \code{data frame} object to identify the column class from select variables of a data frame}

\item{vars}{a \code{character string} or \code{character vector} of column names in the data frame to examine}
}
\value{
Returns a \code{data frame} with two columns where each row is a variable from the data frame of interest
and the data type ()
}
\description{
This \code{map} wrapper function vectorizes the variable data type values within a tidy data frame.
Maps the \code{\link{class}} function for all select variables of a data frame.
}
\examples{
type_df(iris)

}
