% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplotr}
\alias{barplotr}
\title{Bar plot}
\usage{
barplotr(node, Property, y_vari = NULL, MEAN = NULL,
  Interactive = NULL)
}
\arguments{
\item{node}{A data frame}

\item{Property}{A categorical column}

\item{MEAN}{Calculated average of numerical column known as y}

\item{Interactive}{An interactive html plot. Specificy TRUE or FALSE}

\item{y}{A numerical column.}
}
\value{
Barplots of Count, MEAN
}
\description{
Creates barplots for count or mean with SE that can be interactive or not.
}
\examples{





}
