% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse}
\alias{browse}
\title{Browse Dataset}
\usage{
browse(x)
}
\arguments{
\item{x}{The dataset to be opened in the Viewer.}
}
\description{
Browse through your dataset by searching variable names and labels via the search field.
Browser shows variable details such as names, labels, range, distinct values, class, typeof, non-missing cases and missing cases.
}
\examples{
# Either call browse() on a data.frame directly...
browse(mtcars)

# ...or assign and open with View():
df <- browse(mtcars)
# View(df)
}
\keyword{browse}
\keyword{browser}
\keyword{variable}
\keyword{variables}
