% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse}
\alias{browse}
\title{Browse Dataset}
\usage{
browse(x)
}
\arguments{
\item{x}{The dataset to be opened in the Viewer.}
}
\description{
Browse through your dataset by searching variable names and labels via the search field.
Browser shows variable details such as names, labels, range, distinct values, class, typeof, non-missing cases and missing cases.
}
\examples{
# create data.frame
numeric <- c(1, 2, 2, 3, 3, 3, 4, 4, 4, 4)
labelled.numeric <- c(1, 2, 2, 3, 3, 3, 4, 4, 4, 4)
attr(labelled.numeric, "label") <- "my.variable.label"
df <- data.frame(numeric, labelled.numeric)

# open data.frame in Viewer window
}
\keyword{browse}
\keyword{browser}
\keyword{variable}
\keyword{variables}
