% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose}
\alias{decompose}
\title{Get decomposing nodes}
\usage{
decompose(
  attribute,
  node = NULL,
  decomp_type = c("symbol", "scale"),
  scale = NULL,
  symbol = NULL,
  link = NULL,
  filter_fun = NULL
)
}
\arguments{
\item{attribute}{Any node attribute (as a character)}

\item{node}{The MTG node}

\item{decomp_type}{Match the node scale or symbol for decomposition ? If symbol, the function will return all descendants
until matching the same symbol than the node, if scale it will do the same until matching the same scale.}

\item{scale}{An integer vector for filtering the \code{.scale} of the decomposing nodes (i.e. the SCALE
column from the MTG classes).}

\item{symbol}{A character vector for filtering the decomposing nodes by the name of their \code{.symbol} (i.e. the SYMBOL
column from the MTG classes).}

\item{link}{A character vector for filtering the \code{.link} with the descendant.}

\item{filter_fun}{Any filtering function taking a node as input.}
}
\value{
The attribute values from the nodes decomposing the input node
}
\description{
Get attribute values from the nodes that are decomposing a node (e.g. the growing
units decomposing an axis).
}
\details{
This function is mainly used to summarize attributes at a higher scale when they were measured
at a lower scale. For example we can think of an mtg where the length was measured at the internode scale only,
so this function can be used to summarize it at e.g. axis scale. The filters are used to get the nodes of interest only.
For example an axis can be made of internodes and leaves, but the length of an axis is made from the cumulative
length of the internodes only (not the leaves). A second example would be to summarize the total leaf area per axis, where
we would only want to use the area of the leaves (not the internodes).
}
\note{
This function is mainly intended to be used with \code{\link[=mutate_mtg]{mutate_mtg()}}. In this case,
the \code{node} argument can be left empty (or you can put \code{node = node} equivalently).
}
\examples{
filepath= system.file("extdata", "simple_plant_3.mtg", package = "XploRer")
MTG = read_mtg(filepath)

decompose(".symbol", node = extract_node(MTG,"node_3"), decomp_type = "symbol")

# using filters to remove nodes that we don't need:
decompose(".symbol", node = extract_node(MTG,"node_3"), symbol = "Internode",
          decomp_type = "symbol")

# We can check the function worked properly by printing the symbols of each nodes:
print(MTG$MTG, ".symbol")

}
