% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_MTG.R
\name{expand_node}
\alias{expand_node}
\title{Expand MTG line}
\usage{
expand_node(x)
}
\arguments{
\item{x}{A split MTG line (e.g. c("/P1","/A1"))}
}
\value{
A split MTG line with explicitly all nodes
}
\description{
Expand the elements denoted by the syntactic sugar "<<", "<.<", "++" or "+.+"
}
\examples{
x = strsplit(x = "/A1+U85/U86<U87<.<U93<U94<.<U96<U97+.+U100",
"(?<=.)(?=[</+])",perl = TRUE)[[1]]
expand_node(x)

}
\keyword{internal}
