% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtg_helpers.R
\name{get_parent_value}
\alias{get_parent_value}
\title{Get parent value}
\usage{
get_parent_value(..., node, .scale = NULL)
}
\arguments{
\item{...}{Any node attribute (as a character)}

\item{node}{The node (do not put something when used from \code{\link[=mutate_mtg]{mutate_mtg()}})}

\item{.scale}{The names of the MTG scale required (i.e. the SYMBOL from the MTG classes)}
}
\value{
The values of the attributes for the parent of the node
}
\description{
Get the value of one or several variables from the parent node. This is a helper
function that is usually used as input to \code{\link[=mutate_mtg]{mutate_mtg()}} to get the values of the parent node
for all nodes.
}
\details{
This function returns the values of any attribute of the parent of a node. It is
mainly intended to be used in a call to \code{\link[=mutate_mtg]{mutate_mtg()}} (see \code{\link[=mutate_mtg]{mutate_mtg()}} doc for examples).
}
\examples{
filepath= system.file("extdata", "simple_plant.mtg", package = "XploRer")
MTG = read_MTG(filepath)
get_parent_value("Length","Width",  node = FindNode(MTG$MTG, "node_5"))
}
