% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_mtg.R
\name{mutate_mtg}
\alias{mutate_mtg}
\title{Mutate MTG}
\usage{
mutate_mtg(data, ...)
}
\arguments{
\item{data}{A \code{mtg}, as returned by \code{\link[=read_MTG]{read_MTG()}}.}

\item{...}{Name-value pairs of expressions, each with length 1. To access a variable from the mtg,
use \code{node$var} instead of \code{var}. \

The name of each argument will be the name of a new variable, and the value will be its corresponding value. New variables overwrite
existing variables of the same name. The arguments in \code{...} are automatically quoted and evaluated
in the context of the \code{mtg}. They support unquoting and splicing. See the chapter about
\href{https://adv-r.hadley.nz/metaprogramming.html}{metaprogramming} in the book "Advanced R" from H. Wickham
for an introduction to these concepts.}
}
\value{
The MTG invisibly. It is returned for piping purposes only, but it is modified in-place
inside de function.
}
\description{
Adds new variables to an MTG. New variables overwrite existing variables of
the same name.
}
\note{
The function was designed to be used as for \code{\link[dplyr:mutate]{dplyr::mutate()}}, so it uses non-standard
evaluation (NSE). It also returns the \code{mtg} so it can be used with pipes.
}
\examples{
# Import the MTG:
filepath= system.file("extdata", "simple_plant.mtg", package = "XploRer")
MTG = read_MTG(filepath)

# And mutate it by adding two new variables, Length2 and Length3:
mutate_mtg(MTG, Length2 = node$Length + 2, Length3 = node$Length2 * 2)

# note two thing here:
# 1/ We use "node$" to access the values of a variable inside the mtg;
# 2/ Length3 uses the results of Length2 before it even exist. This is because
# The variables are constructed sequentially.

# We can also use pipes:
\dontrun{
read_MTG(filepath)\%>\%
  mutate_mtg(Length2 = node$Length + 2)\%>\%
  plot(.)
}

mutate_mtg(MTG, Width_parent = get_parent_value("Width",  node = node))

ToDataFrameTree(MTG$MTG,"Length","Length2","Length3")

}
