% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_MTG.R
\name{parse_MTG_node_attr}
\alias{parse_MTG_node_attr}
\title{Parse MTG node attributes}
\usage{
parse_MTG_node_attr(node_data, features, attr_column_start, line = NULL)
}
\arguments{
\item{features}{The features data.frame}

\item{attr_column_start}{The index of the column of the first attribute}

\item{MTG_line}{An MTG line (e.g. "/P1/A1")}
}
\value{
A list of attributes
}
\description{
Parse the attributes names, values and type
}
\keyword{internal}
