% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see.R
\name{see}
\alias{see}
\alias{codebook}
\title{Display Variable Details (Stata’s ‘codebook’ on Steroids)}
\usage{
see(x)

codebook(x)
}
\arguments{
\item{x}{The object to be displayed in the console.}
}
\description{
This function displays the following details of variable:
\itemize{
 \item{attributes (i.e. useful to inspect labels)}
 \item{class, typeof, mode, storage.mode}
 \item{table (frequency and percentage of values)}
 \item{summary statistics}
}
See is an S3 generic and currently works with objects of class 'character', 'factor', 'numeric', 'data.frame' and 'labelled'.
}
\examples{
numeric <- c(1, 2, 2, 3, 3, 3, 4, 4, 4, 4)
see(numeric)

labelled.numeric <- c(1, 2, 2, 3, 3, 3, 4, 4, 4, 4)
attributes(labelled.numeric)$label <- "my_variable_label"
codebook(labelled.numeric)
}
\keyword{codebook}
\keyword{see}
