% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{tabler}
\alias{tabler}
\title{Table}
\usage{
tabler(node, ..., y)
}
\arguments{
\item{node}{A dataframe}

\item{...}{A single property or multiple properties including project_id}

\item{y}{A numerical column. Specify FALSE or node$columnname}
}
\description{
Table
}
