% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological_order.R
\name{topological_order}
\alias{topological_order}
\title{Topological order}
\usage{
topological_order(MTG)
}
\arguments{
\item{MTG}{An MTG tree as from \code{\link[=read_MTG]{read_MTG()}}}
}
\value{
Nothing, update the values of the MTG in-place. The MTG is enriched with the
topological order of each node
}
\description{
Compute the topological order of a tree
}
\examples{
filepath= system.file("extdata", "simple_plant.mtg", package = "XploRer")
MTG= read_MTG(filepath)
topological_order(MTG)
print(MTG$MTG, "topological_order")
}
