% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XploRer-package.R
\docType{package}
\name{XploRer-package}
\alias{XploRer}
\alias{XploRer-package}
\title{XploRer: Read, Write, Plot and Analyze MTG (Mutli-scale Tree Graph) files}
\description{
A package to manage plant topology data in the form of MTG.
}
\author{
\strong{Maintainer}: Remi Vezy \email{remi.vezy@cirad.fr} (\href{https://orcid.org/0000-0002-0808-1461}{ORCID})

}
\keyword{internal}
