% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XploRer-package.R
\docType{package}
\name{XploRer-package}
\alias{XploRer}
\alias{XploRer-package}
\title{XploRer: Read, Write, Plot and Analyze MTG (Mutli-Scale Tree Graph) Files}
\description{
Manage plant topology data from MTG files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://vezy.github.io/XploRer/}
  \item \url{https://github.com/VEZY/XploRer}
}

}
\author{
\strong{Maintainer}: Remi Vezy \email{remi.vezy@cirad.fr} (\href{https://orcid.org/0000-0002-0808-1461}{ORCID})

}
\keyword{internal}
