% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XploRer-package.R
\docType{package}
\name{XploRer-package}
\alias{XploRer}
\alias{XploRer-package}
\title{XploRer: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
